requirements.txt
beautifulsoup4==4.11.1
dateparser==1.1.4
fake-useragent==1.1.0
Flask==1.1.4
Flask-Compress==1.4.0
Flask-SQLAlchemy==2.3.2
ipython==8.7.0
logparser==0.8.2
lxml==4.9.1
numpy==1.23.5
pandas==1.5.2
requests==2.25.1
requests-html==0.10.0
Scrapy==2.7.1
scrapy-rotating-proxies==0.6.2
scrapy-user-agents==0.1.1
scrapyd==1.3.0
scrapydweb==1.4.0
scrapyrt==0.13.0
slackclient==2.9.4
SQLAlchemy==1.3.2
user-agents==2.2.0
google-api-core=2.8.2
google-auth=2.11.0
google-cloud=0.34.0
google-cloud-core=2.3.2
google-cloud-pubsub=2.13.11
nested-lookup==0.2.22